@@ -28,9 +28,8 @@ keywords = [
 ]
 
 [tool.poetry.dependencies]
-python = "^3.9"
+python = "^3.11"
 pandas = "^1.5.0"
-pydantic = "^1.10.2"
 loguru = "^0.6.0"
 Faker = "^15.1.1"
 lxml = "^4.9.1"
@@ -44,6 +43,7 @@ html5lib = "^1.1"
 requests = "^2.28.1"
 bs4 = "^0.0.1"
 strenum = "^0.4.10"
+pydantic = "^2.4.2"
 
 [tool.poetry.dev-dependencies]
 pytest = "^7.2.0"
